rendering imgs

+++ src/main/java/com/gottig/portfolio/dto/dtomodel/TechnologyDTO.java
@@ -16,7 +16,7 @@ public class TechnologyDTO {
     
     private String techDescription;
     
-    private Double techLevel;
+    private int techLevel;
     
     private int techIndex;
